Lots of progress
Restructured the "About Me" section, added my top 50 games in there. I then went to the primary page, and added all of the videos for the entries besides 2 (the two being The Red Rain Case and Memories of Closure). I also write beginning blurbs for a good few of the games.